添加自定义QML插件

+++ ros-qt/build-demo02-Desktop-Debug/CMakeFiles/clean_additional.cmake
@@ -1,10 +1,13 @@
 # Additional clean files
+cmake_minimum_required(VERSION 3.16)
 
-file(REMOVE_RECURSE
+if("${CONFIG}" STREQUAL "" OR "${CONFIG}" STREQUAL "Debug")
+  file(REMOVE_RECURSE
   "CMakeFiles/demo02_autogen.dir/AutogenUsed.txt"
   "CMakeFiles/demo02_autogen.dir/ParseCache.txt"
   "demo02_autogen"
   "thread/CMakeFiles/worker_autogen.dir/AutogenUsed.txt"
   "thread/CMakeFiles/worker_autogen.dir/ParseCache.txt"
   "thread/worker_autogen"
-)
+  )
+endif()
